apps/rag-service/pyproject.toml
[project]
name = "aurixa-rag-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "loguru>=0.7.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=3.3.0",
    "rank-bm25>=0.2.0",
    "numpy>=2.0.0",
    "aurixa-db>=0.1.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]
